修改首页显示和详情显示

+++ shopping-service/shopping-api/src/main/java/com/mall/shopping/IProductService.java
@@ -19,7 +19,7 @@ public interface IProductService {
      * @return
      */
     //ProductDetailResponse getProductDetail(ProductDetailRequest request);
-    ProductDetailResponse getProductDetails(Integer id);
+    ProductDetailResponse getProductDetails(Long id);
 
     /**
      * 查询所有商品（分页）
